Don't rename artifacts of projects that use overlays Previously, the archive name of any project that used an overlay was changed to "partial.[extension]". I suppose that was done to indicate that the artifact only includes resources from the base project, NOT the overlay project(s).
But there's a problem with this. Suppose that projects "childA" and "childB" both use a gretty overlay. Project "parent" depends on the two child projects and creates a WAR file that includes their artifacts. However, both of the child artifacts will be named "partial.jar". We now have duplicate artifact entries in our WAR file (WEB-INF/lib/partial.jar)! That's not a problem immediately because WAR files are allowed to include duplicate entries, but what about when we explode the WAR? One of the partial.jar artifacts will be overwritten!

This commit eliminates the artifact renaming.