Fix failure of "./gradlew tasks"
The task couldn't complete due to:
  Could not get unknown property 'generatePomFileForMavenJavaPublication' for task set.
  at [...](/Users/cwardgar/dev/projects/gretty/libs/common.gradle:107)

The offending code in common.gradle was:
  artifactoryPublish {
    dependsOn { project.tasks.generatePomFileForMavenJavaPublication }
    ...
  }

I believe we needed that due to: https://www.jfrog.com/jira/browse/GAP-174.
However, as stated in the thread, the issue was fixed in
'org.jfrog.buildinfo:build-info-extractor-gradle:3.1.1'.
So, we've updated to that version and now we can remove the line.